Refactor: 체험 조회 서버 컴포넌트로 수정
