+++ src/actions/activity.action.ts
@@ -65,28 +65,6 @@ export const updateActivity = async (activityId: number, activityData: PostActiv
   }
 };
 
-// 체험 상세 조회
-export const getActivityById = async (activityId: number) => {
-  try {
-    const response = await fetch(`${API_URL}/activities/${activityId}`, {
-      headers: {
-        "Content-Type": "application/json",
-      },
-      next: { tags: ["activity"] },
-    });
-
-    if (!response.ok) {
-      throw new Error("내 체험 조회 실패");
-    }
-
-    const data = await response.json();
-    console.log("체험 상세 데이터", data);
-    return data;
-  } catch (error) {
-    console.error("체험 데이터 조회 중 오류 발생:", error);
-  }
-};
-
 // 체험 삭제
 export const deleteActivity = async (activityId: number) => {
   try {
@@ -130,7 +108,5 @@ export const uploadActivityImage = async (formData: FormData) => {
   }
 
   const data = await response.json();
-  console.log("Upload successful:", data);
-
   return data.activityImageUrl;
 };
